How to install cuda on Ubuntu 22.04 Requires NVIDIA-510 package to be installed
How to install cuda on Ubuntu 22.04 Requires NVIDIA-510 package to be installed. You will need an account on nvidia developer network to get the cuda to install. 

This should allow machine learning capability and ffmpeg ecoding on the video card. 

Use this at your own risk, it does affect the video on your system, so a lot could happen! You have been warned.